+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   - Create
   - Carousel
 * `ButtonToggle` no longer inherits from the label decorator as it was providing more functionality than required.
+* `Menu` has been updated to use a `<nav>` tag as its root element.
 * `Rainbow` has been updated to no longer use the `react-highcharts` component. To use this component you need to ensure to make the `Highcharts` library available to your application globally.
 
 ## Google Analytics